feat: add cascade synthetic data generation
